Add support to define product quantities for order
Currently when an order fixture is created all products have the
quantity 1.
This change adds ability to define the quantities for added products
under the "fixture/quote_products/quantities" path in the Yaml file.
The format is an indexes list where the index refers to the product
being under the "fixture/quote_products/quote_products" path, and
the value is the quantity.

For example:
fixture:
    model: sales/quote
    attributes:
        store_id: 1
    quote_products:
        products:
            0: 1234
            1: 4567
        quantities:
            0: 10       # Add 10 of product ID 1234 to the shopping basket
            1: 20       # Add 20 of product ID 4567 to the shopping basket